+++ kernel/term.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include <stddef.h>
 
 #include "../drivers/vga.h"
 
@@ -16,8 +17,8 @@ void term_cursor_show();
 
 void term_color(unsigned char color);
 
-int term_putc(char c);
+size_t term_putc(char c);
 
-int term_print(const char * str);
+size_t term_print(const char * str);
 
 #endif // TERM_H
