+++ libc/stdio.h
@@ -2,14 +2,15 @@
 #define STDIO_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
 void itoa(int n, char * str);
 
-int puts(const char * str);
-int putc(char c);
-int puti(int num, int base, bool upper);
-int putu(unsigned int num, unsigned int base, bool upper);
+size_t puts(const char * str);
+size_t putc(char c);
+size_t puti(int num, int base, bool upper);
+size_t putu(unsigned int num, unsigned int base, bool upper);
 
-int printf(const char * fmt, ...);
+size_t printf(const char * fmt, ...);
 
 #endif // STDIO_H
